Solve for dec-5.1
Create Set with all rules. When a number in the update is encountered,
check if there are any pages it must come before. If so, check previous
parts of the update against it to test if a rule has been violated.
If no violation is found, the update is good.